fix find_vstudio sorting/choice of MSVC/VS by forcing numerical sorting
.# Discussion

When multiple MSVC/VS versions are installed and the version is not specified,
`xmake` chooses the most recent version of MSVC/VS. This choice is made based
upon a sort of the `vsver` values (used for registry lookup; some are years
and some are version numbers). Because of the year/version disparity, some
MSVC/VS installation combinations would result in earlier versions being chosen
erroneously. But a numerical (non-string) sort does result in the correct
ordering. This commit forces that numerical-type sort.